# src/main/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: Courtside
  version: 0.1.0
  summary: Court booking for a single sports club.
  description: |
    Courtside is single-tenant: this document describes the API of one club's own instance.
    There is no tenant identifier anywhere in it, and there never will be.

    This document is the source of truth for the API, not a description generated from an
    implementation. An operation that is not here is not part of the API, and a change here is a
    change to a published contract.

    ## Errors

    Every failure is an RFC 9457 problem detail. `type` is always a `urn:courtside:error:<slug>`
    URN: a stable identifier, not a fetchable page. A club runs its instance under its own domain,
    so an `https://` type would name a host nobody operates.

    Two arrays carry everything a client has to show a member, and both hold the same kind of
    entry:

    - `violations` — a rule the request broke. Always an array, even where only one entry is ever
      possible, so a client has one shape to render rather than two.
    - `fieldErrors` — the same, plus the name of the input it came from.

    Neither carries rendered text. Each entry has a `code` naming a message and `params` holding
    the values that message interpolates, and the client resolves the pair against its own message
    catalogue. This is what makes the API usable in a language the server does not ship.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: |
        A server-side session, established by `POST /api/session`. Deliberately not a JWT: the
        club's own frontend and the API share an origin, and an admin must be able to end a
        session immediately, which a self-contained token cannot offer.
    csrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: |
        Required on every state-changing request. Any `GET` issues a readable `XSRF-TOKEN` cookie;
        echo its value in this header. Modelled as a security scheme rather than as a header
        parameter because it is consumed before the operation runs and is never part of its input.

        A write without it answers `401`, not `403`: for an unauthenticated caller the missing
        token is indistinguishable from a missing session.
  schemas:
    DayOfWeek:
      type: string
      enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY]
...
    CancelScope:
      type: string
      description: Which occurrences of a series an operation applies to.
      enum: [THIS, THIS_AND_FOLLOWING, WHOLE_SERIES]
...
    Violation:
      type: object
      description: |
        A rule the request broke, named rather than rendered. The client resolves `code` against
        its own message catalogue and interpolates `params`.
      required: [code, params]
      properties:
        code:
          type: string
          examples: ["booking.rule.advanceWindow.exceeded"]
        params:
          type: object
          additionalProperties: true
          examples: [{ maxDays: 14 }]
    FieldError:
      type: object
      description: A violation that also names the input it came from.
      required: [field, code, params]
      properties:
        field:
          type: string
          examples: ["endsAt"]
        code:
          type: string
          examples: ["validation.ChronologicalSlot"]
        params:
          type: object
          additionalProperties: true
    Problem:
      type: object
      description: RFC 9457 problem detail.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: A `urn:courtside:error:<slug>` URN. Stable, and not a fetchable page.
          examples: ["urn:courtside:error:court-not-bookable"]
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
          description: |
            English, for a developer reading a log. Never shown to a member, and never built from
            a caller's input — the translatable part is in `violations`.
        instance:
          type: string
          format: uri
        violations:
          type: array
          items: { $ref: "#/components/schemas/Violation" }
        fieldErrors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
...
    ActiveRequest:
      type: object
      required: [active]
      properties:
        active: { type: boolean }
    MembershipType:
      type: object
      required: [id, name, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        ruleSetId: { type: [string, "null"], format: uuid }
        active: { type: boolean }
    MembershipTypeRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 60, pattern: ".*\\S.*" }
        ruleSetId: { type: [string, "null"], format: uuid }
...
    SeriesRuleRequest:
      type: object
      description: |
        A recurrence, ending either on a date or after a number of occurrences — exactly one of
        `endsOn` and `occurrenceCount` is given.
      required: [courtIds, cardId, startsOn, startTime, durationMinutes, intervalWeeks, weekdays]
      properties:
        courtIds:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, format: uuid }
        cardId: { type: string, format: uuid }
        startsOn: { type: string, format: date }
        startTime: { type: string, format: time }
        durationMinutes: { type: integer, minimum: 1, maximum: 1440 }
        intervalWeeks: { type: integer, minimum: 1, maximum: 52 }
        weekdays:
          type: array
          minItems: 1
          uniqueItems: true
          items: { $ref: "#/components/schemas/DayOfWeek" }
        endsOn:
          type: [string, "null"]
          format: date
          description: Not before `startsOn`. Mutually exclusive with `occurrenceCount`.
        occurrenceCount:
          type: [integer, "null"]
          minimum: 1
          maximum: 200
          description: Mutually exclusive with `endsOn`.
        note: { type: [string, "null"] }
    CreateSeriesRequest:
      description: |
        A recurrence together with the occurrences to actually write. Previewing needs only the
        recurrence; creating needs the caller's confirmation as well, which is why the two
        operations do not share one shape.
      allOf:
        - $ref: "#/components/schemas/SeriesRuleRequest"
        - type: object
          required: [confirmedStarts]
          properties:
            confirmedStarts:
              type: array
              minItems: 1
              uniqueItems: true
              description: |
                Each entry must be one the preview actually offered — the creator decides which
                occurrences to keep, and nothing is written that they did not confirm.
              items: { type: string, format: date-time }
    Occurrence:
      type: object
      required: [startsAt, endsAt, blockedCourtIds, violations, creatable]
      properties:
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        blockedCourtIds:
          type: array
          items: { type: string, format: uuid }
        violations:
          type: array
          items: { $ref: "#/components/schemas/Violation" }
        creatable: { type: boolean }
    SeriesPreview:
      type: object
      required: [occurrences, creatableCount, truncatedByHorizon]
      properties:
        occurrences:
          type: array
          items: { $ref: "#/components/schemas/Occurrence" }
        creatableCount: { type: integer, format: int64 }
        truncatedByHorizon:
          type: boolean
          description: |
            True when the recurrence would run past the instance's horizon. Said here rather than
            on creation, because by then a dropped date was never among the confirmed occurrences
            to report against.
        horizonLimit: { type: [string, "null"], format: date }
    SeriesCreated:
      type: object
      required: [bookingIds, skipped]
      properties:
        seriesId:
          type: [string, "null"]
          format: uuid
          description: Null when nothing could be created; no empty series is left behind.
        bookingIds:
          type: array
          items: { type: string, format: uuid }
        skipped:
          type: array
          description: Occurrences a conflict or a rule kept from being created.
          items: { type: string, format: date-time }
    MoveRequest:
      type: object
      description: A move must change at least one of the time, the duration or the courts.
      required: [fromBookingId, scope]
      properties:
        fromBookingId: { type: string, format: uuid }
        scope: { $ref: "#/components/schemas/CancelScope" }
        newStartTime: { type: [string, "null"], format: time }
        newDurationMinutes: { type: [integer, "null"], minimum: 1 }
        newCourtIds:
          type: [array, "null"]
          minItems: 1
          uniqueItems: true
          description: Absent leaves the courts as they are; empty is not a request that means anything.
          items: { type: string, format: uuid }
    Move:
      type: object
      required: [bookingId, fromStartsAt, toStartsAt, blockedCourtIds, unbookableCourtIds, violations, executable]
      properties:
        bookingId: { type: string, format: uuid }
        fromStartsAt: { type: string, format: date-time }
        toStartsAt: { type: string, format: date-time }
        blockedCourtIds:
          type: array
          items: { type: string, format: uuid }
        unbookableCourtIds:
          type: array
          items: { type: string, format: uuid }
        violations:
          type: array
          items: { $ref: "#/components/schemas/Violation" }
        executable: { type: boolean }
    MovePreview:
      type: object
      required: [moves, executable]
      properties:
        moves:
          type: array
          items: { $ref: "#/components/schemas/Move" }
        executable: { type: boolean }
    MoveExecuted:
      type: object
      required: [moved]
      properties:
        moved: { type: integer }
...
  responses:
    Unauthenticated:
      description: No session, or a state-changing request without a CSRF token.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Authenticated, but not allowed to do this.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: No such thing.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Invalid:
      description: The request was understood and rejected.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
paths:
...
  /api/booking-series/preview:
    post:
      tags: [booking-series]
      operationId: previewSeries
      summary: What a recurrence would produce, without writing anything.
      description: |
        Mandatory before creating. Open to any authenticated caller, including one who could not
        actually place the card — the card's role requirement is enforced when the series is
        created, exactly as it is for a single booking.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SeriesRuleRequest" }
      responses:
        "200":
          description: Every occurrence the rule produces, with what would block it.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SeriesPreview" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/booking-series:
    post:
      tags: [booking-series]
      operationId: createSeries
      summary: Create the occurrences the caller confirmed.
      description: |
        `confirmedStarts` names exactly the occurrences to create, and each must be one the
        preview offered. What cannot be created is reported as skipped rather than failing the
        request — a single blocked date does not undo the other twenty-five.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSeriesRequest" }
      responses:
        "201":
          description: Created, with whatever was skipped.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SeriesCreated" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/booking-series/{id}:
    delete:
      tags: [booking-series]
      operationId: cancelSeries
      summary: Cancel occurrences of a series.
      description: Calendar semantics — this one, this and all following, or the whole series.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: fromBookingId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: scope
          in: query
          required: true
          schema: { $ref: "#/components/schemas/CancelScope" }
      responses:
        "204":
          description: Cancelled.
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/booking-series/{id}/move/preview:
    post:
      tags: [booking-series]
      operationId: previewSeriesMove
      summary: Where a move would land, and what would block it.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MoveRequest" }
      responses:
        "200":
          description: One entry per occurrence the scope covers.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MovePreview" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/booking-series/{id}/move:
    post:
      tags: [booking-series]
      operationId: moveSeries
      summary: Move occurrences of a series.
      description: All or nothing — if any occurrence cannot move, none does.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MoveRequest" }
      responses:
        "200":
          description: Moved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MoveExecuted" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Some occurrences cannot move; none was moved.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
  /api/admin/membership-types:
    get:
      tags: [admin-membership-types]
      operationId: listMembershipTypes
      summary: Every membershiptype, active or not.
      responses:
        "200":
          description: The list.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MembershipType" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [admin-membership-types]
      operationId: createMembershipType
      summary: Create a membershiptype.
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MembershipTypeRequest" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipType" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: A membership type already carries that name.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/admin/membership-types/{id}:
    get:
      tags: [admin-membership-types]
      operationId: getMembershipType
      summary: One membershiptype.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The membershiptype.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipType" }
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [admin-membership-types]
      operationId: changeMembershipType
      summary: Change a membershiptype.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MembershipTypeRequest" }
      responses:
        "200":
          description: Changed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipType" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A membership type already carries that name.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /api/admin/membership-types/{id}/active:
    put:
      tags: [admin-membership-types]
      operationId: setMembershipTypeActive
      summary: Activate or deactivate a membershiptype.
      description: Deactivating leaves members on it; it only stops new assignment.
      security:
        - sessionCookie: []
          csrfToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActiveRequest" }
      responses:
        "200":
          description: Changed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MembershipType" }
        "400":
          $ref: "#/components/responses/Invalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...